Door System Rewrite
Rewritten the Door System to be more modular and following the same
format as the new Round system. Also fixed a few bugs related to the
round system change. This may be buggy.

Should this version be unplayable, refer to the
Master-Changes_nzTableFallback branch which has all the same gameplay
features but in the old formatting as we rewrite Master-Changes.