Experimental Sprint Fix
This should fix the sprint GetMaxRunSpeed client side bug, by instead
getting the runspeed directly from the player class if it is not
available from GetMaxRunSpeed(). This may or may not cause higher usage,
I do not know if baseclass.Get and player_manager.GetPlayerClass are
heavy functions to run constantly.