Handy Little Function
- Adds a handy little function for making Map Scripts: "lmne()"
If passed with no arguments, will list all entities with names.
If passed with 1 argument, will return the first entity by that name.
If passed with true in the second argument, will list all entities
containing argument 1 in their name and return the first one.
If passed with true in the third argument, will return the entire table
of entities rather than the first one.